requirements.txt
###### Requirements without Version Specifiers ######
pyyaml
pandas
numpy
matplotlib
xlsxwriter